backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
anthropic==0.46.0
google-cloud-storage==2.9.0